% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.barplot.bar}
\alias{add.barplot.bar}
\title{Add Barplot Bar (auto-stacked)}
\usage{
add.barplot.bar(
  data,
  sd = NULL,
  x = 1,
  width = 0.6,
  col = NULL,
  pal = ocean.haline,
  rev.pal = F,
  border = NA,
  angle = NULL,
  density = NULL
)
}
\arguments{
\item{sd}{the uncertainty of data (absolute units)}

\item{x}{the x position to plot the bar}

\item{width}{the width of the bars drawn}

\item{col}{the colors for the bars, should be same length as data}

\item{pal}{the color pallete used in get.pal() if col is not defined}

\item{rev.pal}{boolean, reverse color assignment?}

\item{border}{option passed to rect() controlling border appearance}

\item{angle}{the angle of the hash-marks if density is given}

\item{density}{the density of hash-marks on shape.}

\item{dataa}{vector or numeric for the height of the bar to draw}
}
\description{
Add Barplot Bar (auto-stacked)
}
\author{
Thomas Bryce Kelly
}
