% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.boot.conf}
\alias{add.boot.conf}
\title{Add Bootstrapped Confidence Intervals}
\usage{
add.boot.conf(
  model,
  new.x = c(-1000:1000),
  col = "#55555540",
  conf = c(0.025, 0.975),
  border = FALSE,
  trendline = FALSE
)
}
\arguments{
\item{model}{A bootstrap object (i.e. a dataframe) containing bootstrapped estimates of m and b.}

\item{new.x}{The x values for which predictions are required.}

\item{col}{The desired color of the confidence band. We recomend colors with transparency for plotting.}

\item{conf}{The quantile ranges for the plotting (default is 95% two-tail).}

\item{border}{Do you want a border on the shaded confidence interval?}

\item{trendline}{Should the maximum liklihood values be plotted as a trendline?}
}
\description{
Add confidence bands to a figure based on results of a bootstrap.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
\keyword{Uncertainty}
