% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.boot.trendline.sig}
\alias{add.boot.trendline.sig}
\title{trendline}
\usage{
add.boot.trendline.sig(
  model,
  new.x,
  p = 0.01,
  sig = 0,
  lty = 2,
  lwd = 1,
  col = "black"
)
}
\description{
trendline
}
\author{
Thomas Bryce Kelly
}
