% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.boxplot.box}
\alias{add.boxplot.box}
\title{Add Boxplot Box}
\usage{
add.boxplot.box(
  x,
  y,
  side = 1,
  col = "#33333330",
  border.col = "black",
  line.col = "black",
  point.col = "black",
  width = 0.7,
  lty = 1,
  lwd = 1,
  xlwd = NULL,
  outliers = T,
  pch = 1,
  cex = 1
)
}
\arguments{
\item{x}{The x locations for the boxs}

\item{y}{The y values that will be plotted.}
}
\description{
Add a box to a boxplot that was initialized by plot.boxplot.
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
