% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.colorbar}
\alias{add.colorbar}
\title{Add Colorbar}
\usage{
add.colorbar(min, max, labels = NULL, ticks = NULL,
  pal = "ocean.algae", rev = FALSE, units = "", col.high = "",
  col.low = "", log = FALSE, base = 10, x.pos = 0.875,
  width = 0.05, y.pos = 0.5, height = 0.6, cex = 1,
  cex.units = 1, n = 255)
}
\description{
Add Colorbar
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
