% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{add.contour}
\alias{add.contour}
\title{Add Contour}
\usage{
add.contour(section, field = "Sigma", text.x = NULL, levels = c(25.7,
  26, 26.3, 26.6), col = "black", text.col = "black", lty = 1,
  lwd = 1, f = 2/3, text.default.pos = 0.1)
}
\description{
Add Contour
}
\author{
Thomas Bryce Kelly
}
