% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.plotting.R
\name{add.error.bars.logx}
\alias{add.error.bars.logx}
\title{Add Error Bars (log x)}
\usage{
add.error.bars.logx(x, s.x, y, s.y, base, col = "black")
}
\arguments{
\item{x}{The x values.}

\item{s.x}{The standard deviation or uncertainty of the x values.}

\item{y}{The y values.}

\item{s.y}{The standard deviation or uncertainty of the y values.}

\item{base}{The base of the log transformation on the x-axis.}

\item{col}{The color of the error bars to be drawn.}
}
\description{
A function to add error bars given mean and stdev when the x-axis is log based.
}
\author{
Thomas Bryce Kelly
}
\keyword{Log}
\keyword{Plotting}
