% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.error.bars}
\alias{add.error.bars}
\title{Add Error Bars}
\usage{
add.error.bars(x, s.x, y, s.y, col = "black")
}
\arguments{
\item{x}{The x values.}

\item{s.x}{The standard deviation or uncertainty of the x values.}

\item{y}{The y values.}

\item{s.y}{The standard deviation or uncertainty of the y values.}

\item{col}{The color of the error bars to be drawn.}
}
\description{
A function to add error bars to a figure.
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
