% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.exaggerated.axis}
\alias{add.exaggerated.axis}
\title{Exaggerate (Transform) Data for plotting}
\usage{
add.exaggerated.axis(
  side,
  power = 0.8,
  at = NULL,
  labels = NULL,
  grid = F,
  lty = 2,
  col = "grey",
  ...
)
}
\arguments{
\item{power}{a positive value, typically less than 1, which is the exponent of the transformation (default = 0.8)}

\item{...}{optional values that are passed onto axis()}

\item{x}{data values to be transformed (e.g. depth)}
}
\description{
Exaggerate (Transform) Data for plotting
}
\author{
Thomas Bryce Kelly
}
