% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{add.lm.conf}
\alias{add.lm.conf}
\title{Add Linear Model Confidence Intervals}
\usage{
add.lm.conf(model, x = NULL, name = NULL, col = "#50505030", level = 0.95, ...)
}
\description{
A helper function to plot the confidence intervals determined from a base::lm model.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
