% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{add.lm.trendline}
\alias{add.lm.trendline}
\title{Add Linear Model Trendline}
\usage{
add.lm.trendline(model, col = "black", lwd = 1, lty = 1, ...)
}
\description{
A helper function to plot the trendline determined from a base::lm model.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
