% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.arrows}
\alias{add.map.arrows}
\title{Add Map Arrows}
\usage{
add.map.arrows(map, lon, lat, lon2, lat2, col = "black", cex = 1, lwd = 1, ...)
}
\arguments{
\item{...}{optional arguments passed to shape::Arrows().}
}
\description{
Add station points to a map
}
\author{
Thomas Bryce Kelly
}
