% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.bathy}
\alias{add.map.bathy}
\title{Add Map Bathymetry (shaded)}
\usage{
add.map.bathy(
  map,
  bathy,
  refine = 0,
  trim = T,
  zlim = NULL,
  pal = "greyscale",
  n = 255,
  rev = FALSE,
  col.low = "",
  col.high = "",
  verbose = T
)
}
\arguments{
\item{map}{a map object returned by the function make.map()}

\item{bathy}{a bathymetric dataframe returned by get.map.bathy}

\item{zlim}{the z-axis range that the colormap is projected upon}

\item{pal}{the color palette used for coloring the shading. Should consist of a function name as a string}

\item{n}{the number of distinct colors to use}

\item{rev}{a boolean flag to flip the color axis}

\item{col.low}{the color used when painting data that is below the zlim range. A value of NA causes the data not to be drawn, and a value of '' uses the lowest color value for out of range data.}

\item{col.high}{same as \emph{col.low} but for data that is out of range above the zlim.}

\item{filled}{a flag to turn on color smoothing and filling from the OCE package. Works for some datasets but may cause issues.}

\item{refinement}{the level of bilinear refinement to apply to the image grid}
}
\description{
Add Map Bathymetry (shaded)
}
\author{
Thomas Bryce Kelly
}
