% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{add.map.bathy.shade}
\alias{add.map.bathy.shade}
\title{Add Map Bathymetry (shaded)}
\usage{
add.map.bathy.shade(map, bathy, pal = "ocean.ice", n = 255,
  zlim = NULL, rev = FALSE, filled = TRUE, col.low = NA,
  col.high = NA)
}
\arguments{
\item{map}{a map object returned by the function make.map()}

\item{bathy}{a bathymetric dataframe returned by get.map.bathy}

\item{pal}{the color palette used for coloring the shading. Should consist of a function name as a string}

\item{n}{the number of distinct colors to use}

\item{zlim}{the z-axis range that the colormap is projected upon}

\item{rev}{a boolean flag to flip the color axis}

\item{filled}{a flag to turn on color smoothing and filling from the OCE package. Works for some datasets but may cause issues.}

\item{col.low}{the color used when painting data that is below the zlim range. A value of NA causes the data not to be drawn, and a value of '' uses the lowest color value for out of range data.}

\item{col.high}{same as \emph{col.low} but for data that is out of range above the zlim.}
}
\description{
Add Map Bathymetry (shaded)
}
\author{
Thomas Bryce Kelly
}
