% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.coastline}
\alias{add.map.coastline}
\title{Add Coastline to Map}
\usage{
add.map.coastline(coast, p, land.col, field.inv = NULL)
}
\description{
Adds a coastline list to a map given a projection and land color.
}
\author{
Thomas Bryce Kelly
}
