% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.contour}
\alias{add.map.contour}
\title{Add Map Contours}
\usage{
add.map.contour(
  map,
  lon,
  lat,
  z,
  trim = T,
  col = "black",
  levels = NULL,
  zlim = NULL,
  n = 3,
  labels = TRUE,
  lty = 1,
  lwd = 1,
  verbose = T
)
}
\description{
Add Map Contours
}
\author{
Thomas Bryce Kelly
}
