% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.grid}
\alias{add.map.grid}
\title{Add Grid to Map}
\usage{
add.map.grid(map, lons = NULL, lats = NULL, col = "darkgrey", lty = 2, lwd = 1)
}
\description{
Add Grid to Map
}
\author{
Thomas Bryce Kelly
}
