% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.points}
\alias{add.map.points}
\title{Add Map Points}
\usage{
add.map.points(map, lon, lat, col = "black", cex = 1, pch = 16, ...)
}
\arguments{
\item{lon}{the longitudes of the points to be drawn}

\item{lat}{the latitude of the points to be drawn}

\item{col}{the color of the points}

\item{cex}{the size of point to be drawn}

\item{pch}{the point character to be used}

\item{...}{optional arguments passed to points().}

\item{stn.lon}{(depreciated) the longitudes of the points to be drawn}

\item{stn.lat}{(depreciated) the latitude of the points to be drawn}
}
\description{
Add station points to a map
}
\author{
Laura Whitmore
}
