% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{add.map.polygon}
\alias{add.map.polygon}
\title{Add Map Polygon}
\usage{
add.map.polygon(
  lon,
  lat,
  col = "#00000020",
  lty = 1,
  lwd = 1,
  border = NULL,
  density = NULL,
  angle = 45,
  fillOddEven = FALSE
)
}
\description{
Add Map Polygon
}
\author{
Laura Whitmore
}
