% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.quiver}
\alias{add.map.quiver}
\title{Add Quiver Lines}
\usage{
add.map.quiver(
  map,
  lon,
  lat,
  u,
  v,
  zscale = 1,
  col = "black",
  lwd = 1,
  verbose = T
)
}
\description{
Add Quiver Lines
}
\author{
Thomas Bryce Kelly
}
