% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{add.map.scale}
\alias{add.map.scale}
\title{Add Map Scalebar}
\usage{
add.map.scale(pos = 1, scale = NULL, verbose = T, col = "black", cex = 1)
}
\description{
Add Map Scalebar
}
\author{
Thomas Bryce Kelly
}
