% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{add.map.text}
\alias{add.map.text}
\title{Add Map Text}
\usage{
add.map.text(lon, lat, text, col = "black", cex = 1, adj = NULL, pos = NULL)
}
\arguments{
\item{lon}{longitude position of the text or a vector of positions}

\item{lat}{latitude position of the text or a vector of positions}

\item{text}{a string or vector of strings for the text to be written}

\item{col}{text color}

\item{cex}{size of the text to be written}

\item{adj}{one or two values in [0, 1] which specify the x (and optionally y) adjustment of the labels.}

\item{pos}{a position specifier for the text. If specified this overrides any adj value given. Values of 1, 2, 3 and 4, respectively indicate positions below, to the left of, above and to the right of the specified coordinates.}
}
\description{
Add Map Text
}
\author{
Thomas Bryce Kelly

Laura Whitmore
}
