% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{add.regress.trendline}
\alias{add.regress.trendline}
\title{Add Bootstrapped Trendline}
\usage{
add.regress.trendline(model, col = "black", lty = 2, lwd = 1, ...)
}
\description{
Add the maximum likelihood trendline to a figured based on a bootstrap estimation.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
