% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{add.section.bathy}
\alias{add.section.bathy}
\title{Add Section Bathymetry}
\usage{
add.section.bathy(section, stn.lon, stn.lat, stn.depth = NULL,
  by.lat = TRUE, bathy.col = "darkgrey", res = 2)
}
\description{
Add Section Bathymetry
}
\author{
Thomas Bryce Kelly
}
