% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.plot.R
\name{add.section.inlay}
\alias{add.section.inlay}
\title{Add Map Inlay}
\usage{
add.section.inlay(
  section,
  p = NULL,
  side = 1,
  pch = 20,
  cex = 1,
  col = "red",
  land.col = "black",
  height = 0.1,
  width = 0.1
)
}
\description{
Add Map Inlay
}
