% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{advect}
\alias{advect}
\title{Advect Lagrangian Particle}
\usage{
advect(
  lon,
  lat,
  depth,
  time,
  advection,
  dt,
  get.vel,
  sink = NULL,
  const.depth = NULL
)
}
\arguments{
\item{lon}{Initial longitude of particle}

\item{lat}{Initial latitude of particle}

\item{depth}{Intial depth of particle}

\item{time}{Start time of particle}

\item{advection}{An advection list, such as one generated by `load.oscar.advection()`, that contains U V and W}

\item{dt}{Time step for RK4 advection scheme}
}
\description{
Advect Lagrangian Particle
}
\author{
Thomas Bryce Kelly
}
