% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bathy.global}
\alias{bathy.global}
\title{Global Bathymetry

A Matrix containing elevation data taken from NOAA via marmap.}
\format{A matrix containing 2160 x 1080 values of elevation. Depth is given in negative. Resolution is 10 arc minutes (1/6th degree).}
\source{
NOAA Bathymetric Database
}
\usage{
bathy.global
}
\description{
Global Bathymetry

A Matrix containing elevation data taken from NOAA via marmap.
}
\keyword{datasets}
