% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bathy.pacific}
\alias{bathy.pacific}
\title{Pacific Bathymetry

A Matrix containing elevation data taken from NOAA via marmap.}
\format{
A list
}
\source{
NOAA Bathymetric Database
}
\usage{
bathy.pacific
}
\description{
Pacific Bathymetry

A Matrix containing elevation data taken from NOAA via marmap.
}
\keyword{datasets}
