% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(x, s.x, y, s.y, n = 1000)
}
\arguments{
\item{x}{The observed x values}

\item{s.x}{The uncertainty in each x observation}

\item{y}{The observed y values}

\item{s.y}{The uncertainty in each y observation}

\item{n}{The number of bootstrapped models to generate.}
}
\description{
Generate a bootstrapped model from observations with known uncertainty.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
