% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.misc.R
\name{build.impute}
\alias{build.impute}
\title{Impute Missing Values}
\usage{
build.impute(x, method = "forest")
}
\arguments{
\item{x}{the data frame with missing values}

\item{method}{the imputation method: 'mean', 'median', 'forest}
}
\description{
Impute Missing Values
}
