% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{build.section.multi.old}
\alias{build.section.multi.old}
\title{Build Section (multi, old)}
\usage{
build.section.multi.old(x, y, z, xlim = NULL, ylim = NULL,
  x.factor = 1, y.factor = 1, x.scale = NULL, y.scale = NULL,
  uncertainty = 0, neighborhood = 20, field.names = NULL)
}
\description{
Build Section (multi, old)
}
\author{
Thomas Bryce Kelly
}
