% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{build.section.multi}
\alias{build.section.multi}
\title{Build Section (multi)}
\usage{
build.section.multi(x, y, z, xlim = NULL, ylim = NULL, x.factor = 1,
  y.factor = 1, x.scale = NULL, y.scale = NULL,
  uncertainty = 1e-08, neighborhood = 20, field.names = NULL)
}
\description{
Build Section (multi)
}
\author{
Thomas Bryce Kelly
}
