% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{build.section.odv}
\alias{build.section.odv}
\title{Build Section (ODV)}
\usage{
build.section.odv(x, y, z, xlim = NULL, ylim = NULL, x.factor = 1,
  y.factor = 1, x.scale = NULL, y.scale = NULL, uncertainty = 0,
  neighborhood = 20, field.name = "z")
}
\arguments{
\item{x}{Dimensions (e.g. lat, lon, depth, section distance, time, etc)}

\item{y}{Dimensions (e.g. lat, lon, depth, section distance, time, etc)}

\item{z}{signal to be gridded (e.g. T, S, ...)}

\item{xlim}{Limits of the gridding. These are the bounds of the new x-y grid. Default: NULL will set it based on the data + 10%.}

\item{x.factor, }{y.factor The relative scale difference between x and y, used to calculate distances. Take into account actual scale AND the relevent scaling of the system (vertical distance tends to be more important than horizontal distance).}

\item{x.scale, }{y.scale The step size in the new x-y grid. By default the scale is set to generate a grid that is 50x50.}

\item{uncertainty}{Scaling applied to the distance from the cener of a grid cell to a vertex, used to add a base-line distance to all measurements. 0 = no minimum, 1 = minimum = to half a grid cell.}

\item{neighborhood}{The number of closest points used in the interpolation of a grid cell. Special value of -1 specifies the use of all points. If neighborhood > number of valid data then neighborhood is set to equal the number of valid datapoints.}

\item{field.name}{Sets the name of the new interpolated field. By default the name is 'z', but setting it to 'T' for temperature makes sense.}
}
\description{
Build Section (ODV)
}
\author{
Thomas Bryce Kelly
}
