% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{build.section.old}
\alias{build.section.old}
\title{Build Section (old)}
\usage{
build.section.old(x, y, z, xlim = NULL, ylim = NULL, x.factor = 1,
  y.factor = 1, x.scale = NULL, y.scale = NULL,
  uncertainty = 1e-12, neighborhood = -1, field.name = "z")
}
\description{
Build Section (old)
}
\author{
Thomas Bryce Kelly
}
