% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{build.section}
\alias{build.section}
\title{Build Section}
\usage{
build.section(
  x,
  y,
  z,
  lat = NULL,
  lon = NULL,
  xlim = NULL,
  ylim = NULL,
  x.factor = 1,
  y.factor = 1,
  x.scale = NULL,
  y.scale = NULL,
  uncertainty = 1e-12,
  p = 3,
  gridder = gridIDW,
  field.names = NULL,
  nx = 50,
  ny = 50,
  by.lat = FALSE
)
}
\description{
Build Section
}
\author{
Thomas Bryce Kelly
}
\keyword{Gridding}
