% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.adiabatic.temp.grad}
\alias{calc.adiabatic.temp.grad}
\title{Calculate Adiabatic Temperature Gradient}
\usage{
calc.adiabatic.temp.grad(S, T, P, verbose = T)
}
\arguments{
\item{S}{Salinity in PSU}

\item{T}{Temperature in centigrade}

\item{P}{Pressure in db}
}
\description{
Calculate Adiabatic Temperature Gradient
}
\author{
Thomas Bryce Kelly
}
