% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.AOU}
\alias{calc.AOU}
\title{Calculate AOU}
\usage{
calc.AOU(T, S, P, oxy)
}
\arguments{
\item{T}{temperature}

\item{S}{salinity}

\item{P}{pressure}

\item{oxy}{oxygen in umol kg-1}
}
\description{
Calculate AOU
}
