% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.ArSol}
\alias{calc.ArSol}
\title{Calculate Argon Solubility}
\usage{
calc.ArSol(S = 35, T = 10, verbose = T)
}
\arguments{
\item{S}{Salinity in PSU}

\item{T}{Temperature}
}
\description{
Calculate Argon solubility in seawater. Should return a value of 13.46 in 10 degree water in 35 PSU water.
}
\author{
Thomas Bryce Kelly
}
