% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{calc.boot.ss}
\alias{calc.boot.ss}
\title{Calculate Bootstrapped Sum of Squares}
\usage{
calc.boot.ss(y)
}
\description{
A helper function for calculating SSR from a series of bootstraps.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
