% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.freezing.point}
\alias{calc.freezing.point}
\title{Calculate Freezing Point in Seawater}
\usage{
calc.freezing.point(S, p, f = 0, verbose = T)
}
\description{
Calculate Freezing Point in Seawater
}
