% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.k}
\alias{calc.k}
\title{Calculate k}
\usage{
calc.k(u, SST, verbose = T)
}
\description{
Calculate k (aka piston velocity) for a given windspeed and water temperature.
}
\references{
Wanninkhof, R. (2014). Relationship between wind speed and gas exchange over the ocean revisited: Gas exchange and wind speed over the ocean. Limnology and Oceanography: Methods 12, 351–362. doi:10.4319/lom.2014.12.351.

Wanninkhof, R. (1992). Relationship between wind speed and gas exchange over the ocean. Journal of Geophysical Research 97, 7373. doi:10.1029/92JC00188.

Sweeney, C., Gloor, E., Jacobson, A. R., Key, R. M., McKinley, G., Sarmiento, J. L., et al. (2007). Constraining global air-sea gas exchange for CO 2 with recent bomb 14 C measurements: BOMB 14 C AND AIR-SEA GAS EXCHANGE. Global Biogeochemical Cycles 21, n/a-n/a. doi:10.1029/2006GB002784.
}
