% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.k.reuer}
\alias{calc.k.reuer}
\title{Calculate k (Reuer)}
\usage{
calc.k.reuer(ws, mld, SST, dt, teeter.mod = TRUE)
}
\arguments{
\item{ws}{A vector of prior windspeeds ordered from most recent to oldest.}

\item{mld}{Mixed layer depth in meters}

\item{SST}{Mixed layer temperature in C}

\item{dt}{a numeric or vector of the temporal spacing of the wind speed observations in days}

\item{teeter.mod}{A boolean flag to turn on or off the modified version of Reuer's weighting scheme as outlined in Teeter et al. 2018.}
}
\value{
It will return a list with the individual weights and weighted sum (for dianostics), the k value (that you want), the instantaneous k (also useful), k.hist showing the history of k values (diagnostic), and time is time before measurement that each diagnostic is referenced against (for diagnostics).
}
\description{
This function calculates the temporal length scale of the NCP measurement. Essentially it calculates the ventilation based on water column and wind speed over a sequence of measurements (equally spaced).
}
\references{
Reuer, M. K., Barnett, B. A., Bender, M. L., Falkowski, P. G., and Hendricks, M. B. (2007). New estimates of Southern Ocean biological production rates from O2/Ar ratios and the triple isotope composition of O2. Deep Sea Research Part I: Oceanographic Research Papers 54, 951–974. doi:10.1016/j.dsr.2007.02.007.
}
\author{
Thomas Bryce Kelly
}
\keyword{Air-sea}
\keyword{NCP}
\keyword{Oceanography}
