% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.mld}
\alias{calc.mld}
\title{Calculate MLD}
\usage{
calc.mld(depth, density, set.depth = 10, delta = 0.1, resolution = 0.1)
}
\arguments{
\item{depth}{depth}

\item{density}{the density of the seawater. Alternatively, if using a temperature based MLD definition, the temperature of the seawater.}

\item{set.depth}{the depth from which the mixed layer character is defined}

\item{delta}{the difference from the mixed layer signal that defines the end of the mixed layer.}

\item{resolution}{the resolution of the vertical interpolation for when calculating the MLD.}
}
\value{
returns the depth where the signal has changed from set.depth by >= delta.
}
\description{
Calculate MLD
}
\author{
Thomas Bryce Kelly
}
