% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.N2Sol}
\alias{calc.N2Sol}
\title{Calculate N2 Solubility}
\usage{
calc.N2Sol(S = 35, Tmp = 10, verbose = T)
}
\arguments{
\item{S}{Salinity (numeric or a vector) in PSU}

\item{Tmp}{Temperature (numeric or a vector) in centigrade}
}
\description{
Calculate the N2 solubility in seawater.
}
\author{
Thomas Bryce Kelly
}
