% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.NeSol}
\alias{calc.NeSol}
\title{Calculate Neon Solubility}
\usage{
calc.NeSol(S = 35, T = 10)
}
\arguments{
\item{S}{Salinity in PSU}

\item{T}{Temperature in centigrade}
}
\description{
Calculates the Neon solubility in seawater based on salinity and temperature
}
\references{
Hamme & Emerseon 2004 (Table 4)
}
\author{
Thomas Bryce Kelly
}
