% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.O2sol}
\alias{calc.O2sol}
\title{Calcualte O2 Solubility}
\usage{
calc.O2sol(S = 35, T = 10)
}
\arguments{
\item{S}{Salinity in PSU}

\item{T}{Temperature in centigrade}
}
\description{
Calcualte O2 Solubility
}
\author{
Thomas Bryce Kelly
}
