% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.POstar}
\alias{calc.POstar}
\title{Calculate Phosphate Star (PO4Star)}
\usage{
calc.POstar(Oxy, PO4, R = 1/175, verbose = T)
}
\description{
Calculate Phosphate Star (PO4Star)
}
\references{
broeker and peng 1998
}
