% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.pref.PO4}
\alias{calc.pref.PO4}
\title{Calculate Preformed Phosphate}
\usage{
calc.pref.PO4(AOU, PO4, R = 1/175)
}
\description{
Calculate Preformed Phosphate
}
\references{
broeker and peng 1998
}
