% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{calc.r.squared}
\alias{calc.r.squared}
\title{Calculate r-squared}
\usage{
calc.r.squared(x, y)
}
\arguments{
\item{x}{A set of data (e.g. observations).}

\item{y}{A set of data (e.g. model output).}
}
\description{
A function to calculate the coefficient of determination between paired data. Note: calc.r.squared(x,y) = calc.r.squared(y,x).
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
