% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.rho}
\alias{calc.rho}
\title{Calculate Density}
\usage{
calc.rho(S = 30, T = 15, p = 0)
}
\arguments{
\item{S}{Salinity in PSU}

\item{T}{Temperature in centigrade}

\item{p}{Pressure in db}
}
\description{
Calculate Density
}
\references{
Massel 2015
}
\author{
Thomas Bryce Kelly
}
