% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.schmidt.number}
\alias{calc.schmidt.number}
\title{Calculate Schmidt Number}
\usage{
calc.schmidt.number(SST, verbose = T)
}
\arguments{
\item{SST}{Water temperature in degree centigrade}
}
\description{
Calcualtes the schmidt number for seawater at a given temperature.
}
\references{
Wanninkhof, R. (2014). Relationship between wind speed and gas exchange over the ocean revisited: Gas exchange and wind speed over the ocean. Limnology and Oceanography: Methods 12, 351–362. doi:10.4319/lom.2014.12.351.
}
\author{
Thomas Bryce Kelly
}
