% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.seawater.compressibility}
\alias{calc.seawater.compressibility}
\title{Calculate Seawater Compressibility (module)}
\usage{
calc.seawater.compressibility(S = 8, T = 10, P = 10, verbose = T)
}
\arguments{
\item{P}{Pressure in bar}
}
\description{
Calculate Seawater Compressibility (module)
}
\references{
https://link.springer.com/content/pdf/bbm%3A978-3-319-18908-6%2F1.pdf
}
\author{
Thomas Bryce Kelly
}
