% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.sigma.theta}
\alias{calc.sigma.theta}
\title{Calculate Potential Density Anomaly}
\usage{
calc.sigma.theta(S, T, P, P.ref = 0, verbose = T)
}
\arguments{
\item{S}{Salinity in PSU}

\item{T}{Temperature in centigrade}

\item{P}{Sample pressure in db}

\item{P.ref}{Reference pressure, typically 0 db (surface).}
}
\description{
Function to calculate potential density anomaly from T, S and pressure. Comonly called sigma-theta, potential density anomaly is a useful water mass tracer due to conservation of potential density due to adiabatic processes.
}
\examples{
calc.sigma.theta(S = 35, T = 8, P = 350)
}
\author{
Thomas Bryce Kelly
}
