% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.misc.R
\name{calc.sma}
\alias{calc.sma}
\title{Calculate Simple Moving Average}
\usage{
calc.sma(parameter, k = 100)
}
\arguments{
\item{parameter}{signal or vector to be smoothed}

\item{k}{the integer number of entries to average over}
}
\description{
Calculate Simple Moving Average
}
\author{
Laura Whitmore
}
