% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coastline3}
\alias{coastline3}
\title{World Coastline Data (L3)
This is a coarse resolution coastline, compiled from the global hierarchical coastline data product.}
\format{
An object of class \code{list} of length 2.
}
\source{
NOAA
}
\usage{
coastline3
}
\description{
World Coastline Data (L3)
This is a coarse resolution coastline, compiled from the global hierarchical coastline data product.
}
\keyword{datasets}
