% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coastline5}
\alias{coastline5}
\title{World Coastline Data (L5)
This is a coarse resolution coastline, compiled from the global hierarchical coastline data product.}
\format{
An object of class \code{list} of length 2.
}
\source{

}
\usage{
coastline5
}
\description{
World Coastline Data (L5)
This is a coarse resolution coastline, compiled from the global hierarchical coastline data product.
}
\keyword{datasets}
