% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coastlineWorld}
\alias{coastlineWorld}
\title{World Coastline Data
This is a coarse resolution coastline, copied from the OCE package, at scale 1:110M, with 10,696 points, suitable for world-scale plots plotted at a small size, e.g. inset diagrams.}
\format{
An object of class \code{coastline} of length 1.
}
\source{
Downloaded from https://www.naturalearthdata.com, in ne_110m_admin_0_countries.shp in July 2015, with an update on December 16, 2017. Copied from the OCE R package.
}
\usage{
coastlineWorld
}
\description{
World Coastline Data
This is a coarse resolution coastline, copied from the OCE package, at scale 1:110M, with 10,696 points, suitable for world-scale plots plotted at a small size, e.g. inset diagrams.
}
\keyword{datasets}
