% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.misc.R
\name{conv.kahru.npp}
\alias{conv.kahru.npp}
\title{Convert Data to NPP (Kahru)}
\usage{
conv.kahru.npp(x)
}
\arguments{
\item{x}{raw value}
}
\description{
Converts kahru's data to npp
}
\author{
Thomas Bryce Kelly
}
