% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.misc.R
\name{conv.kahru.sst}
\alias{conv.kahru.sst}
\title{Convert Data to SST (Kahru)}
\usage{
conv.kahru.sst(x)
}
\arguments{
\item{x}{The pixel values (or a vector or matrix)}
}
\description{
Convert
}
\author{
Thomas Bryce Kelly
}
