% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{conv.p.to.d}
\alias{conv.p.to.d}
\title{Convert Pressure to Depth}
\usage{
conv.p.to.d(
  p = 10000,
  latitude = 30,
  geo.anom = 0,
  d = NULL,
  rev = FALSE,
  verbose = T
)
}
\arguments{
\item{p}{Pressure in db}

\item{latitude}{Latitude in degrees N}

\item{geo.anom}{Geopotential anomaly in meters}

\item{d}{Depth in meters, NULL unless using the function in reverse}

\item{rev}{Run the function in reverse? (i.e. Depth -> Pressure)}
}
\description{
Convert between pressure and depth using UNESCO 1983 coefficients.
}
\references{
UNESCO 1983
}
\author{
Thomas Bryce Kelly
}
