% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{conv.time.excel}
\alias{conv.time.excel}
\title{Convert Time Excel}
\usage{
conv.time.excel(x, tz = "GMT", rev = FALSE)
}
\arguments{
\item{x}{The object to be converted. Typically a numeric when converting from excel to a POSIXct.}

\item{tz}{The timezone of the epoch used to save the numeric type, almost \emph{ALWAYS} GMT.}

\item{rev}{Whether or not the operation should be reversed. This is a loseless operation.}
}
\description{
Helper function for converting the datetime stamps from Microsoft Excel.
}
\author{
Thomas Bryce Kelly
}
