% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{conv.time.matlab}
\alias{conv.time.matlab}
\title{Convert Time Matlab}
\usage{
conv.time.matlab(x, tz = "GMT")
}
\arguments{
\item{x}{The numeric timestamp that will be converted.}
}
\description{
Convert Time Matlab
}
