% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.roms.r
\name{conv.time.roms}
\alias{conv.time.roms}
\title{Convert ROMS Time}
\usage{
conv.time.roms(x, tz = "UTC")
}
\arguments{
\item{x}{The datetime numeric to be converted}

\item{tz}{The timezone for the conversion, almost always UTC}
}
\description{
Convert ROMS Time
}
\author{
Thomas Bryce Kelly
}
