% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{conv.time.unix}
\alias{conv.time.unix}
\title{Convert Time Unix}
\usage{
conv.time.unix(x, tz = "GMT")
}
\arguments{
\item{x}{The numeric value that will be converted.}

\item{tz}{The timezone of the epoch used for the timestamp, almost always GMT.}
}
\description{
A function used to convert a unix timestamp into a POSIXct object.

Typical examples of unix timestamps include any POSIX object that has been coerced into a numeric.
}
