% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{conv.uM.to.ml}
\alias{conv.uM.to.ml}
\title{Convert Oxygen Units}
\usage{
conv.uM.to.ml(x, rev = FALSE, verbose = FALSE)
}
\description{
Convert Oxygen Units
}
\references{
https://ocean.ices.dk/tools/unitconversion.aspx
}
\author{
Thomas Bryce Kelly
}
