% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{conv.watt.to.einstein}
\alias{conv.watt.to.einstein}
\title{Convert Light Units}
\usage{
conv.watt.to.einstein(x, verbose = T)
}
\description{
(6.02e23 quanta/Ein) / (86,400 seconds/day * 2.77e18 quanta/s/W)
}
\references{
Unit conversions as listed by MBARI: https://www3.mbari.org/bog/nopp/par.html
}
\author{
Thomas Bryce Kelly
}
