% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{data.frame}
\alias{data.frame}
\title{Make dataframe}
\usage{
data.frame(...)
}
\description{
A helpful wrapper to the base data.frame function which provides column indexes upon mismatched lengths.
}
\author{
Thomas Bryce Kelly
}
