% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{eez}
\alias{eez}
\title{World EEZ Data
Global maritime eez boundaries. Creative Commons Attribution 4.0}
\format{
An object of class \code{list} of length 161.
}
\source{
Flanders Marine Institute (2019). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 11. Available online at https://www.marineregions.org/. https://doi.org/10.14284/386
}
\usage{
eez
}
\description{
World EEZ Data
Global maritime eez boundaries. Creative Commons Attribution 4.0
}
\keyword{datasets}
