% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.plotting.R
\name{exaggerate}
\alias{exaggerate}
\title{Exaggerate (Transform) Data for plotting}
\usage{
exaggerate(x, power = 0.8)
}
\arguments{
\item{x}{data values to be transformed (e.g. depth)}

\item{power}{a positive value, typically less than 1, which is the exponent of the transformation (default = 0.8)}
}
\description{
Exaggerate (Transform) Data for plotting
}
\author{
Thomas Bryce Kelly
}
