% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Geotraces.GA01.Sal}
\alias{Geotraces.GA01.Sal}
\title{Geotraces' GA01 Salinity and Temperature Data
A dataframe containing T and S with various metadata.}
\format{A Dataframe containing 5103 observations of 12 variables}
\source{
See the Geotraces Intermediate Data Product and refences at http://www.geotraces.org/dp/idp2017
}
\usage{
Geotraces.GA01.Sal
}
\description{
Geotraces' GA01 Salinity and Temperature Data
A dataframe containing T and S with various metadata.
}
\keyword{datasets}
