% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.bathy.R
\name{get.bathy}
\alias{get.bathy}
\title{Get Bathymetry}
\usage{
get.bathy(map, res = 25, override = FALSE, keep = FALSE)
}
\arguments{
\item{map}{a map object as returned by make.map()}

\item{res}{the resolution of the bathymetery to be requested. Value is in arc minutes. Smaller values will take longer to download and to plot.}

\item{override}{a flag to override the default longitude bounds used in order to facilitate high latitude map making.}

\item{keep}{a flag used to determine if the bathymetry data should be saved to enable offline reuse later.}
}
\description{
Get the bathymetry you want from NOAA (marmap)
}
\author{
Laura Whitmore

Thomas Bryce Kelly
}
