% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{get.boot.vals}
\alias{get.boot.vals}
\title{Get Bootstrapped Values}
\usage{
get.boot.vals(model, x, conf = 0.5)
}
\arguments{
\item{model}{A bootstrap object (i.e. a dataframe) containing bootstrapped estimates of m and b.}

\item{x}{The x values for which predictions are required.}

\item{conf}{The confidence interval of the predicted values saught. A value of 0.5 is the maximum likelihood value.}
}
\description{
Get predicted values from a series of bootstraps.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
