% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{get.coast}
\alias{get.coast}
\title{Load Coastline Data}
\usage{
get.coast(coast = "coastlineWorld")
}
\arguments{
\item{cost}{a string referring to a coastline data object such as those provided by the OCE package.}
}
\description{
Low res to high res: coastlineWorld, coastlineWorldCoarse, coastlinewWorldMedium, coastlineWorldFine
}
\author{
Laura Whitmore

Thomas Bryce Kelly
}
