% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{get.map.line}
\alias{get.map.line}
\title{Add Map Line}
\usage{
get.map.line(map, lon, lat, greatCircle = T)
}
\arguments{
\item{lon}{a set of longitudes to draw a line between}

\item{lat}{a set of latitudes to draw a lien between}

\item{col}{the color of the line to be drawn}

\item{lty}{the type of line to be drawn}

\item{lwd}{the width of the line to be drawn}
}
\description{
Add Map Line
}
\author{
Thomas Bryce Kelly
}
