% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{get.month}
\alias{get.month}
\title{Get Month}
\usage{
get.month(x)
}
\arguments{
\item{x}{An object of class POSIX or a vector containing POSIX.}
}
\description{
A helper function used to pull the our information from a POSIX object.
}
