% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.load.R
\name{get.satellite.times}
\alias{get.satellite.times}
\title{Get Satellite Times}
\usage{
get.satellite.times(x, verbose = T)
}
\arguments{
\item{x}{raw}
}
\description{
time
}
\author{
Thomas Bryce Kelly
}
