% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.plot.R
\name{get.section.bathy}
\alias{get.section.bathy}
\title{Get Section Bathymetry}
\usage{
get.section.bathy(lon, lat, res = 10)
}
\arguments{
\item{lon}{longitudes}

\item{lat}{latitudes}

\item{res}{the resolution of the bathymetry (arc minutes, e.g. 10)}
}
\description{
retreive bathymetry for a section plot
}
\author{
Thomas Bryce Kelly
}
\keyword{Bathymetry}
