% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{get.vel.globcurrent}
\alias{get.vel.globcurrent}
\title{Get Velocity (OSCAR)}
\usage{
get.vel.globcurrent(lon, lat, depth, time, advection)
}
\description{
Get Velocity (OSCAR)
}
\author{
Thomas Bryce Kelly
}
