% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{get.year}
\alias{get.year}
\title{Get Year}
\usage{
get.year(x)
}
\arguments{
\item{x}{An object of class POSIX or a vector containing POSIX.}
}
\description{
A helper function used to pull the our information from a POSIX object.
}
