% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.pal.r
\name{greyscale}
\alias{greyscale}
\title{Greyscale Palette}
\usage{
greyscale(n, rev = FALSE)
}
\arguments{
\item{n}{the number of greyscale colors desired}

\item{rev}{a boolean flag to reverse the color palette}
}
\description{
Greyscale Palette
}
\author{
Thomas Bryce Kelly
}
