% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.misc.R
\name{grid.interp}
\alias{grid.interp}
\title{Interpolate from fractional index onto grid}
\usage{
grid.interp(grid, i, j)
}
\description{
Interpolate from fractional index onto grid
}
\author{
Thomas Bryce Kelly
}
