% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.gridder.R
\name{gridBin}
\alias{gridBin}
\title{Grid via Binning}
\usage{
gridBin(gx, gy, x, y, z, p, xscale, yscale, uncertainty)
}
\arguments{
\item{gx}{Grid x values to interpolate onto}

\item{gy}{Grid y values to interpolate onto}

\item{x}{Observations, x values}

\item{y}{Observations, y values}

\item{z}{Observations, z values}

\item{p}{Unused}

\item{xscale}{The spacing of the x grid}

\item{yscale}{The spacing of the y grid}

\item{uncertainty}{Unused}
}
\description{
Bins data into reguarly spaced grid
}
\author{
Thomas Bryce Kelly
}
