% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.gridder.R
\name{gridIDW}
\alias{gridIDW}
\title{Grid via Inverse Distance Weighting Interpolation}
\usage{
gridIDW(gx, gy, x, y, z, p, xscale, yscale, uncertainty)
}
\arguments{
\item{gx}{Grid x values to interpolate onto}

\item{gy}{Grid y values to interpolate onto}

\item{x}{Observations, x values}

\item{y}{Observations, y values}

\item{z}{Observations, z values}

\item{p}{Exponent on the distance function}
}
\description{
Grid via Inverse Distance Weighting Interpolation
}
\author{
Thomas Bryce Kelly
}
