% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.gridder.R
\name{gridKrig}
\alias{gridKrig}
\title{Grid via Krigging}
\usage{
gridKrig(
  gx,
  gy,
  x,
  y,
  z,
  p = 2,
  xscale = 1,
  yscale = 1,
  uncertainty = 0.1,
  neighborhood = NULL,
  x.factor = 1,
  y.factor = 1
)
}
\arguments{
\item{gx}{Grid x values to interpolate onto}

\item{gy}{Grid y values to interpolate onto}

\item{x}{Observations, x values}

\item{y}{Observations, y values}

\item{z}{Observations, z values}

\item{p}{Unused}
}
\description{
Grid via Krigging
}
\author{
Thomas Bryce Kelly
}
