% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{gridNN.old}
\alias{gridNN.old}
\title{Grid via Nearest Neighbor}
\usage{
gridNN.old(gx, gy, x, y, z, p, xscale, yscale, uncertainty)
}
\description{
Grid via Nearest Neighbor
}
\author{
Thomas Bryce Kelly
}
