% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{gridNNI}
\alias{gridNNI}
\title{Grid via Natural Neighbor Interpolation}
\usage{
gridNNI(gx, gy, x, y, z, p, xscale, yscale, uncertainty)
}
\description{
Grid via Natural Neighbor Interpolation
}
\author{
Thomas Bryce Kelly
}
