% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{init.lagrangian.history}
\alias{init.lagrangian.history}
\title{Initialize Lagrangian History}
\usage{
init.lagrangian.history(meta, extra.tracers = NULL, verbose = T)
}
\arguments{
\item{meta}{list object containing lagrangian model setup (see init.lagrangian.model())}
}
\description{
Initialize Lagrangian History
}
\author{
Thomas Bryce Kelly
}
