% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{init.lagrangian.model}
\alias{init.lagrangian.model}
\title{Initialize Lagrangian Model}
\usage{
init.lagrangian.model(
  t.start,
  t.end,
  t.step,
  save.freq,
  nparticles,
  verbose = T
)
}
\arguments{
\item{t.start}{start time for the Lagrangian model (POSIX)}

\item{t.end}{end time for the Lagrangian model (POSIX)}

\item{t.step}{time between internal model steps (sec)}

\item{save.freq}{number of time steps between particle saves}

\item{nparticles}{maximum number of particles considered}

\item{verbose}{output flag for setup data}
}
\description{
Initialize Lagrangian Model
}
\author{
Thomas Bryce Kelly
}
