% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{init.lagrangian.particles}
\alias{init.lagrangian.particles}
\title{Initiate Lagrangian Particles}
\usage{
init.lagrangian.particles(
  lon = NA,
  lat = NA,
  depth = NA,
  time = NA,
  sink = NA,
  const.depth = NA,
  density = NA,
  verbose = T
)
}
\arguments{
\item{lon}{Longitude values of the lagrangian floats}

\item{lat}{Latitude values of the lagrangian floats}

\item{depth}{Depths of the lagrangian floats}

\item{time}{POSIX datetime of lagrangian floats}

\item{sink}{Sinking velocity (m s-1) of the lagrangian float}

\item{const.depth}{Depth (m) at which to maintain regardless of vertical velocity}

\item{density}{Density (kg m-3) of the float in order to maintain isopycnal position}

\item{verbose}{Verbose flag, default = T}
}
\description{
Initiate Lagrangian Particles
}
\author{
Thomas Bryce Kelly
}
