% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{is.inside}
\alias{is.inside}
\title{Is Inside Polygon}
\usage{
is.inside(pos, box, verbose = FALSE)
}
\description{
Useful to determine if a set of coordinates lies inside or outside a closed polygon. Currently works for
 2D coordinates such as x,y or lat,lon. Example box argument: box = data.frame(x = c(-1, 1, 1, -1), y = c(1, 1, -1, -1)).
}
\author{
Thomas Bryce Kelly
}
