% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{is.nan}
\alias{is.nan}
\title{Is NAN (extension of R base)}
\usage{
is.nan(x)
}
\arguments{
\item{x}{object to test for NAN}
}
\description{
A simple extension to the base is.nan() function to permit operation on data.frames.
}
\author{
Christian Fender
}
