% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{is.POSIXct}
\alias{is.POSIXct}
\title{Is POSIXct}
\usage{
is.POSIXct(x)
}
\description{
A helper function to determine if an object or a vector of objects is POSIX.
}
\author{
Thomas Bryce Kelly
}
\keyword{Helper}
\keyword{Time}
