% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{is.prime}
\alias{is.prime}
\title{Is Prime?}
\usage{
is.prime(n)
}
\arguments{
\item{n}{a numeric or vector of numerics to test if values are prime}
}
\description{
Is Prime?
}
\author{
Thomas Bryce Kelly
}
