% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{jackknife}
\alias{jackknife}
\title{Jackknife}
\usage{
jackknife(x, y, n = 1000)
}
\arguments{
\item{x}{The observed x values}

\item{y}{The observed y values}

\item{n}{The number of jackknife models to generate.}
}
\description{
Generate a bootstrapped model from observations with unknown uncertainty.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
