% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{jackknife.section}
\alias{jackknife.section}
\title{Run Jackknife Analysis on Section}
\usage{
jackknife.section(section, n)
}
\description{
Run Jackknife Analysis on Section
}
\author{
Thomas Bryce Kelly
}
