% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lakes5}
\alias{lakes5}
\title{World lake Polygon (L5)
This is a coarse resolution coastline, compiled from the global hierarchical coastline data product.}
\format{
An object of class \code{list} of length 2.
}
\source{
NOAA
}
\usage{
lakes5
}
\description{
World lake Polygon (L5)
This is a coarse resolution coastline, compiled from the global hierarchical coastline data product.
}
\keyword{datasets}
