% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{load.advection.cmems}
\alias{load.advection.cmems}
\title{Load in CMEMS Advection}
\usage{
load.advection.cmems(
  path,
  lats = NULL,
  lons = NULL,
  get.vel = get.vel.cmems,
  verbose = T
)
}
\arguments{
\item{path}{CMEMS NC file (includign path) to read in}

\item{verbose}{A verbose flag}
}
\description{
Load in CMEMS Advection
}
\author{
Thomas Bryce Kelly
}
