% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{load.advection.ecco}
\alias{load.advection.ecco}
\title{Load in CMEMS Advection}
\usage{
load.advection.ecco(
  u.path,
  v.path,
  w.path,
  lats = NULL,
  lons = NULL,
  get.vel = NULL,
  verbose = T
)
}
\arguments{
\item{verbose}{A verbose flag}

\item{path}{CMEMS NC file (includign path) to read in}
}
\description{
Load in CMEMS Advection
}
\author{
Thomas Bryce Kelly
}
