% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{load.advection.globcurrent}
\alias{load.advection.globcurrent}
\title{Load Advection (globcurrent)}
\usage{
load.advection.globcurrent(
  file,
  lats = NULL,
  lons = NULL,
  get.vel = get.vel.globcurrent,
  verbose = F
)
}
\arguments{
\item{file}{globcurrent file (includign path) to read in}

\item{lats}{The latitude limits of the resulting model domain}

\item{lons}{The longitude of the resulting model domain}

\item{verbose}{A verbose flag}
}
\description{
A function used to load an OSCAR circulation netcdf file and structure in a manner suitable for Lagrangian particle release.
}
\author{
Thomas Bryce Kelly
}
