% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{load.advection.mitgcm}
\alias{load.advection.mitgcm}
\title{Load Advection (MITgcm)}
\usage{
load.advection.mitgcm(u.file, v.file, w.file = NULL)
}
\description{
A function used to load an mitgcm file set (offline data fields) and structure in a manner suitable for Lagrangian particle release.
}
\author{
Thomas Bryce Kelly
}
