% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{load.advection.roms}
\alias{load.advection.roms}
\title{Load in ROMS Advection}
\usage{
load.advection.roms(
  path,
  lats = NULL,
  lons = NULL,
  get.vel = get.vel.roms,
  verbose = T
)
}
\arguments{
\item{path}{ROMS file (includign path) to read in}

\item{verbose}{A verbose flag}
}
\description{
Load in ROMS Advection
}
\author{
Thomas Bryce Kelly
}
