% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.frrf.r
\name{load.frrf}
\alias{load.frrf}
\title{Load FRRF Datafiles}
\usage{
load.frrf(input.dir, file.names)
}
\arguments{
\item{input.dir}{the input directory}

\item{files.names}{the name of the file in the input directory}
}
\description{
Load FRRF Datafiles
}
\author{
Thomas Bryce Kelly
}
