% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.nc.R
\name{load.nc.subset}
\alias{load.nc.subset}
\title{Load NetCDF}
\usage{
load.nc.subset(file, var = NULL, x = NULL, y = NULL, test = F, verbose = T)
}
\arguments{
\item{file}{The location of a .nc file.}
}
\description{
Load NetCDF
}
