% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.satellite.r
\name{load.satellite}
\alias{load.satellite}
\title{Read Satellite Data}
\usage{
load.satellite(
  input.dir = NULL,
  file = NULL,
  conv = function(x) {     x },
  entry = 1
)
}
\arguments{
\item{input.dir}{the input directory}
}
\description{
read
}
\author{
Thomas Bryce Kelly
}
