% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{ma}
\alias{ma}
\title{Moving Average}
\usage{
ma(x, n = 5)
}
\arguments{
\item{x}{A vector containing equally spaced observations}

\item{n}{The number of samples to include in the moving average.}
}
\description{
Calculate the moving average from a series of observations. \emph{NB}: It assumes equally spaced observations
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
