% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{make.cex}
\alias{make.cex}
\title{Calculate point sizes}
\usage{
make.cex(x, min = 0.4, max = 4, log = FALSE, base = 10)
}
\description{
Returns a vector of cex values for use in plotting points with size dependent on a vector of values. For example, plotting a scatter plot where point size is proportional to population size.
}
\author{
Thomas Bryce Kelly
}
