% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{make.map.arctic}
\alias{make.map.arctic}
\title{Make an Arctic Map}
\usage{
make.map.arctic(
  coast = "coastlineWorld",
  lon.min = -180,
  lon.max = 180,
  lat.min = 60,
  lat.max = 90,
  p = make.proj("stere", lat = 90),
  dlat = 15,
  dlon = 60,
  land.col = "lightgray",
  grid = TRUE
)
}
\description{
Make an Arctic Map
}
