% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.test.R
\name{make.map.cce}
\alias{make.map.cce}
\title{Make Map (CCE)}
\usage{
make.map.cce(
  coast = "coastline3",
  lon.min = -126,
  lon.max = -119,
  lat.min = 31,
  lat.max = 36,
  p = make.proj(projection = "merc", lat = 35, lon = -122, dlat = 3),
  land.col = "#252525",
  draw.grid = TRUE,
  dlon = 3,
  dlat = 3
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastline2'.}

\item{lon.min}{The minimum longitude displayed on the map.}

\item{lon.max}{The maximum longitude on the map}

\item{lat.min}{The minimum latitude shown on the map}

\item{lat.max}{The maximum latitude shown on the map}

\item{p}{a projection string, such as those generated by 'make.proj()'}

\item{land.col}{A color string or value for the land polygon}

\item{dlon}{The spacing for the longitude grid (in degrees)}

\item{dlat}{The spacing for the latitude grid (in degrees)}
}
\description{
An example map of the CCE study region.
}
\author{
Thomas Bryce Kelly
}
