% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.test.R
\name{make.map.nga}
\alias{make.map.nga}
\title{Make Map (NGA)}
\usage{
make.map.nga(
  coast = "coastlineWorldFine",
  lon.min = -154,
  lon.max = -149,
  lat.min = 56,
  lat.max = 61,
  p = make.proj("stere", lat = 60, lon = -151),
  dlat = 3,
  dlon = 3,
  land.col = "#333333",
  draw.grid = TRUE
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastlineWorld'.}

\item{lon.min}{The minimum longitude displayed on the map.}

\item{lon.max}{The maximum longitude on the map}

\item{lat.min}{The minimum latitude shown on the map}

\item{lat.max}{The maximum latitude shown on the map}

\item{p}{a projection string, such as those generated by 'make.proj()'}

\item{dlat}{The spacing for the latitude grid (in degrees)}

\item{dlon}{The spacing for the longitude grid (in degrees)}

\item{land.col}{A color string or value for the land polygon}
}
\description{
An example map of the NGA -- basemap only.
}
\author{
Thomas Bryce Kelly
}
