% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{make.map2}
\alias{make.map2}
\title{Make Map}
\usage{
make.map2(
  coast = NULL,
  lon = 0,
  lat = 0,
  scale = 1000,
  p = NULL,
  land.col = "lightgray",
  draw.grid = T,
  draw.axis = T,
  dlon = 60,
  dlat = 20,
  verbose = T
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastlineWorld'.}

\item{p}{a projection string, such as those generated by 'make.proj()'}

\item{land.col}{A color string or value for the land polygon}

\item{dlon}{The spacing for the longitude grid (in degrees)}

\item{dlat}{The spacing for the latitude grid (in degrees)}

\item{lon.min}{The minimum longitude displayed on the map.}

\item{lon.max}{The maximum longitude on the map}

\item{lat.min}{The minimum latitude shown on the map}

\item{lat.max}{The maximum latitude shown on the map}

\item{grid}{Boolean, draw lat/lon grid?}
}
\description{
Default settings are for Arctic, lowest resolution coastline.
Exaples of projection arguments include: `+proj=stere +lat_0=90`, `+proj=merc`, `+proj=aea +lat_1=30 +lat_2=45 +lon_0=-120`.
}
\author{
Laura Whitmore

Thomas Bryce Kelly
}
