% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{make.pal}
\alias{make.pal}
\title{Make Pal}
\usage{
make.pal(
  x,
  n = 255,
  min = NA,
  max = NA,
  pal = "greyscale",
  rev = FALSE,
  clip = FALSE
)
}
\arguments{
\item{x}{the values to be color coded}

\item{n}{the number of distinct colors to use}

\item{min}{the minimum value to correspond to the first color}

\item{max}{the maximum value to correspond to the last color}

\item{pal}{the pallete to use, default is ocean.haline}

\item{rev}{Boolean, reverse the color pallete?}

\item{clip}{boolean, remove out of range values? Defaults to False}
}
\description{
Make Pal
}
\author{
Thomas Bryce Kelly
}
