% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.pal.r
\name{make.qual.pal}
\alias{make.qual.pal}
\title{Make Qualitative Palette}
\usage{
make.qual.pal(x, pal = NULL, rev = FALSE)
}
\arguments{
\item{x}{The categorical values to be colored}

\item{pal}{The palette to be used, default tol}

\item{rev}{Boolean, reverse the colors?}
}
\description{
Make Qualitative Palette
}
\author{
Thomas Bryce Kelly
}
