% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.frrf.r
\name{model.Eilers1988}
\alias{model.Eilers1988}
\title{Eilers 1988 Light Model}
\usage{
model.Eilers1988(alpha, Eopt, Pm, E)
}
\arguments{
\item{alpha}{the initial slope of the P vs E curve}

\item{E}{the photosynthetically active radiation}
}
\description{
Eilers 1988 Light Model
}
