% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.frrf.r
\name{model.Jassby1976}
\alias{model.Jassby1976}
\title{Jassby 1976 Light Model}
\usage{
model.Jassby1976(alpha, Ek, E)
}
\arguments{
\item{alpha}{the initial slope of the P vs E curve}

\item{Ek}{a light saturation parameter}

\item{E}{the photosynthetically active radiation}
}
\description{
Jassby 1976 Light Model
}
