% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frrf.r
\name{model.Platt1980}
\alias{model.Platt1980}
\title{Platt 1980 Light Curve}
\usage{
model.Platt1980(alpha, beta, Ps, E)
}
\arguments{
\item{alpha}{the initial slope of the P vs E curve}

\item{E}{the photosynthetically active radiation}
}
\description{
Platt 1980 Light Curve
}
