% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.frrf.r
\name{model.Webb1974}
\alias{model.Webb1974}
\title{Webb 1974 Light Curve}
\usage{
model.Webb1974(alpha, Ek, E)
}
\arguments{
\item{alpha}{the initial slope of the P vs E curve}

\item{Ek}{a light staturation parameter}

\item{E}{the photosynthetically active radiation}
}
\description{
Webb 1974 Light Curve
}
