% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{parameter.descent}
\alias{parameter.descent}
\title{Parameter Search}
\usage{
parameter.descent(
  cost,
  guess = NULL,
  ...,
  bounds = NULL,
  max.iter = 100,
  progression = NULL,
  step = NULL,
  tol = 1e-08,
  verbose = T
)
}
\arguments{
\item{cost}{The cost function which must return a numeric value and accept parameter values as the first arguments and in the order they are provided.}

\item{...}{Optional argument that is passed directly onto the cost function}

\item{bounds}{A dataframe containing the minimum and maximum values permitted of each parameter}

\item{progression}{The size of the new search-space to interrogate. A value between 1 and splits/2. Default value (NULL) will yeield a progression of max(1, splits/4), good for most problems.}

\item{n}{Number of recursions to perform}

\item{splits}{The number of subdivisions to perform for each dimension (so grid size is n x splits ^ dimensionality)}
}
\description{
Implements a recursive grid search routine to solve optimization problems in arbitrary dimensions.
}
\author{
Thomas Bryce Kelly
}
