% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.optimization.R
\name{parameter.search.bootstrap}
\alias{parameter.search.bootstrap}
\title{Parameter Search with Randomization}
\usage{
parameter.search.bootstrap(n, cost, ..., bounds, splits = 10)
}
\arguments{
\item{n}{Number of recursions to perform}

\item{cost}{The cost function which must return a numeric value and accept parameter values as the first arguments and in the order they are provided.}

\item{...}{Optional argument that is passed directly onto the cost function}

\item{bounds}{A dataframe containing the minimum and maximum values permitted of each parameter}

\item{splits}{The number of subdivisions to perform for each dimension (so grid size is n x splits ^ dimensionality)}
}
\description{
Implements a recursive grid search routine to solve optimization problems in arbitrary dimensions.
}
\author{
Thomas Bryce Kelly
}
