% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.optimization.R
\name{parameter.search.parallel}
\alias{parameter.search.parallel}
\title{Parameter Search for Parallel Processing}
\usage{
parameter.search.parallel(
  n,
  cost,
  grid = NULL,
  ...,
  bounds,
  splits = 10,
  progression = NULL
)
}
\arguments{
\item{n}{Number of recursions to perform}

\item{cost}{The cost function which must return a numeric value and accept parameter values as the first arguments and in the order they are provided.}

\item{...}{Optional argument that is passed directly onto the cost function}

\item{bounds}{A dataframe containing the minimum and maximum values permitted of each parameter}

\item{splits}{The number of subdivisions to perform for each dimension (so grid size is n x splits ^ dimensionality)}

\item{progression}{The size of the new search-space to interrogate. A value between 1 and splits/2. Default value (NULL) will yeield a progression of max(1, splits/4), good for most problems.}
}
\description{
Implements a recursive grid search routine to solve optimization problems in arbitrary dimensions.
}
\author{
Thomas Bryce Kelly
}
