% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{parameter.walk}
\alias{parameter.walk}
\title{Parameter Search using Random Walk}
\usage{
parameter.walk(
  n,
  cost,
  ...,
  bounds,
  progression = NULL,
  max.iter = 1e+06,
  k = 2
)
}
\arguments{
\item{n}{Number of accepted solutions to seek}

\item{cost}{The cost function which must return a numeric value and accept parameter values as the first arguments and in the order they are provided.}

\item{...}{Optional argument that is passed directly onto the cost function}

\item{bounds}{A dataframe containing the minimum and maximum values permitted of each parameter}

\item{progression}{The step length for each parameter (default is 1% of bounded range).}

\item{max.iter}{The maximum number of steps to take before returning (even if n is unsatisfied). **Goal for function to return via n rather than by max.iter.}
}
\description{
Implements an MCMC algorithm (metropolis criterion) to determine a global optimum from within the stated bounds.
}
\author{
Thomas Bryce Kelly
}
