% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.beta.r
\name{plot.beta.counts}
\alias{plot.beta.counts}
\title{Plot Beta Counts}
\usage{
plot.beta.counts(data, ylim = NULL, cols = pals::glasbey(5),
  show.legend = TRUE)
}
\description{
A helper function to quickly plot up raw data from a beta count data file.
}
\author{
Thomas Bryce Kelly
}
\keyword{Thorium}
