% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.boxplot}
\alias{plot.boxplot}
\title{Plot Boxplot}
\usage{
plot.boxplot(
  n = 10,
  ylim = c(0, 1),
  at = NULL,
  main = NULL,
  labels = NULL,
  ylab = "x",
  xlab = "y",
  ...
)
}
\arguments{
\item{n}{The width of the x axis, i.e. n = 10 corresponds to xlim = c(1,10)}

\item{ylim}{the y axis limit}

\item{at}{Values to draw x-axis labels}

\item{main}{The plot's main title}

\item{labels}{The labels to use for the x-axis}

\item{ylab}{The y axis label}

\item{xlab}{The x axis label}

\item{...}{Optional arguemnts passed to plot()}
}
\description{
Generate a blank boxplot
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
