% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.grid}
\alias{plot.grid}
\title{Plot strcutured grid}
\usage{
plot.grid(
  x,
  y,
  z = NULL,
  col = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "x",
  ylab = "y",
  zlim = NULL,
  pal = "greyscale",
  n = 255,
  rev = F
)
}
\description{
Function used to plot grid data (an array) with arbitrary positions, which can be provided as either an array of z values (along with optional palette options) or as an array of colors directly.
}
\author{
Thomas Bryce Kelly
}
