% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.image}
\alias{plot.image}
\title{Plot Image}
\usage{
plot.image(
  x = NULL,
  y = NULL,
  z,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  pal = "greyscale",
  n = 255,
  rev = F,
  ...
)
}
\description{
Plot Image
}
\author{
Thomas Bryce Kelly
}
