% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.plot.R
\name{plot.section}
\alias{plot.section}
\title{Plot Section}
\usage{
plot.section(
  section,
  field = NULL,
  xlim = NULL,
  ylim = NULL,
  by.lon = F,
  by.lat = F,
  xlab = "x",
  ylab = "y",
  log = FALSE,
  base = 10,
  zlim = NULL,
  pal = "greyscale",
  rev = F,
  include.data = F,
  mark.points = F,
  include.pch = 21,
  include.cex = 1,
  main = NULL,
  col.low = "",
  col.high = "",
  N = 255
)
}
\arguments{
\item{field}{Parameter name to be plotted (z-axis).}

\item{xlim}{Limits of the plot, by default the limits are set to the limits of the section grid.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{log}{Flag for turning on log transformation of the z-axis. If TRUE then z = log(z) is performed prior to plotting.}

\item{base}{Used when log = TRUE to set the base of the log.}

\item{zlim}{The zlim imposed on the z-axis, which by default is set to the range of z values in the data.}

\item{pal}{The color palette used; should be modeled after get.pal().}

\item{rev}{Boolean for reversing the palette colors.}

\item{include.data}{Flag for whether the data used to construct the grid should be plotted using the same palette.}

\item{mark.points}{Flag for marking the location of each sample used to construct the grid.}

\item{include.pch}{The pch used for when mark.points = TRUE.}

\item{include.cex}{Sets the point size for when either include.data or mark.points are TRUE. (Used for both).}

\item{main}{The title text to be included in the top line of the plot. Defaults to the name of the field.}

\item{col.low}{[unimplemented] The color used when plotting out-of-range z values on the low end. Default value of '' indicates to use the minimum value of pal(). A value of NA skips the plotting of these data. Otherwise the color given is used (e.g. col.low = 'blue').}

\item{col.high}{[unimplemented] Same as col.low but for out-of-range high values.}

\item{section:}{section object (2D only) to be plotted.}

\item{tlim}{Limits of the plot, by default the limits are set to the limits of the section grid.}
}
\description{
Plot a section object.
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
\keyword{Section}
