% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{plot.TS}
\alias{plot.TS}
\title{Make TS Plot}
\usage{
plot.TS(
  S,
  Tmp,
  xlim = c(25, 36),
  ylim = c(-5, 15),
  levels = seq(0, 40, by = 2),
  cex.lab = 1,
  drawlabels = TRUE,
  labels = NULL,
  freezing.line = T,
  freezing.col = "#00000030",
  col.contour = "grey",
  lwd = 1,
  lty = 1,
  xlab = "Practical Salinity",
  ylab = "Potential Temperature",
  pch = 1,
  col = "black",
  cex = 1,
  main = NULL
)
}
\description{
Make TS Plot
}
\author{
Thomas Bryce Kelly
}
