% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.misc.R
\name{rebuild.satellite.index}
\alias{rebuild.satellite.index}
\title{Rebuild Satellite Index}
\usage{
rebuild.satellite.index(dir = "Z:/Data/Data_Satellite/Raw", verbose = FALSE)
}
\arguments{
\item{verbose}{Flag to print diagnostics}
}
\description{
rebuild
}
\author{
Thomas Bryce Kelly
}
