% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{redraw.map}
\alias{redraw.map}
\title{Redraw Map}
\usage{
redraw.map(map, coast = NULL, land.col = NULL, grid = NULL)
}
\arguments{
\item{map}{a map object as returned by make.map()}
}
\description{
Redraw Map
}
\author{
Thomas Bryce Kelly
}
