% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.regress.R
\name{regress.bootstrap}
\alias{regress.bootstrap}
\title{Bootstrap}
\usage{
regress.bootstrap(x, y, sx, sy, n = 1000)
}
\arguments{
\item{x}{The observed x values}

\item{y}{The observed y values}

\item{sx}{The uncertainty in each x observation}

\item{sy}{The uncertainty in each y observation}

\item{n}{The number of bootstrapped models to generate.}
}
\description{
Generate a bootstrapped model from observations with known uncertainty (and x - y independence).
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
