% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{regress.bootstrap.type2}
\alias{regress.bootstrap.type2}
\title{Bootstrap 2}
\usage{
regress.bootstrap.type2(x, s.x, y, s.y, n = 1000)
}
\description{
Bootstrap 2
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
