% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{regress.jackknife.2}
\alias{regress.jackknife.2}
\title{Jackknife Model II}
\usage{
regress.jackknife.2(x, y, sx = NA, sy = NA, n = 1000)
}
\arguments{
\item{x}{The observed x values}

\item{y}{The observed y values}

\item{n}{The number of jackknife models to generate.}
}
\description{
Generate a bootstrapped model from observations with unknown uncertainty.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
