% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.regress.R
\name{regress.jackknife}
\alias{regress.jackknife}
\title{Jackknife}
\usage{
regress.jackknife(x, y, sx = NA, sy = NA)
}
\arguments{
\item{x}{The observed x values}

\item{y}{The observed y values}

\item{n}{The number of jackknife models to generate.}
}
\description{
Generate a bootstrapped model from observations with unknown uncertainty.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
