% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{replot.map}
\alias{replot.map}
\title{Replot Map}
\usage{
replot.map(map = map)
}
\arguments{
\item{map}{a map object as returned by make.map()}
}
\description{
Replot Map
}
\author{
Thomas Bryce Kelly
}
