% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{run.advection.fast}
\alias{run.advection.fast}
\title{Advect Lagrangian Particles Fast}
\usage{
run.advection.fast(
  particles,
  model,
  advection,
  zlim = c(-6000, 0),
  verbose = T
)
}
\arguments{
\item{particles}{a particle release dataframe such as that generated by `init.lagrangian.particles()`}

\item{model}{a model list such as that generated by `init.lagrangian.model()`}

\item{advection}{An advection list, such as one generated by `load.oscar.advection()`, that contains U V and W}

\item{dt}{Time step for forward euler advection}
}
\description{
Advect Lagrangian Particles Fast
}
\author{
Thomas Bryce Kelly
}
