% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.lagrangian.r
\name{run.advection.parallel}
\alias{run.advection.parallel}
\title{Advect Lagrangian Particle Model in Parallel}
\usage{
run.advection.parallel(
  particles,
  meta,
  hist,
  advection,
  get.vel,
  zlim = c(-6000, 0),
  verbose = T
)
}
\arguments{
\item{advection}{An advection list, such as one generated by `load.oscar.advection()`, that contains U V and W}

\item{lat}{Initial latitude of particle}

\item{lon}{Initial longitude of particle}

\item{depth}{Intial depth of particle}

\item{time}{Start time of particle}

\item{dt}{Time step for RK4 advection scheme}
}
\description{
Advect Lagrangian Particle Model in Parallel
}
\author{
Thomas Bryce Kelly
}
