% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sample.frrf}
\alias{sample.frrf}
\title{Sample FRRF Datafile
A sample of an FRRF datafile for use in the demonstrations and tutorials.}
\format{
a list containing the datastructure returned by TheSource::load.frrf()
}
\source{
Sven A Kranz, see Kranz et al. 2020 in JGR: Oceans.
}
\usage{
sample.frrf
}
\description{
Sample FRRF Datafile
A sample of an FRRF datafile for use in the demonstrations and tutorials.
}
\keyword{datasets}
