% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{sync.dir}
\alias{sync.dir}
\title{Sync file directory}
\usage{
sync.dir(source, destination, pattern = "*", verbose = T, level = 0, sep = "/")
}
\arguments{
\item{source}{String leading to the folder to be synced}

\item{destination}{Destination where files and folders should be copied.}

\item{pattern}{File/folder name value to be matched against possible files and folders. Default is '*'}

\item{verbose}{Output flag}

\item{level}{A recursion counter for the number of sublevels indexed (used for logging)}

\item{sep}{The folder/folder seperator, default is}
}
\description{
Sync file directory
}
