% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.optimization.R
\name{test.function}
\alias{test.function}
\title{Get Optimization Test Function}
\usage{
test.function(n, verbose = T)
}
\arguments{
\item{n}{the number of the test function desired.}
}
\description{
Get Optimization Test Function
}
\references{
Wikipedia: https://en.wikipedia.org/wiki/Test_functions_for_optimization
}
\author{
Thomas Bryce Kelly
}
