% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.test.R
\name{test.map.california}
\alias{test.map.california}
\title{Test Map of California}
\usage{
test.map.california(
  coast = "coastline3",
  p = "+proj=merc",
  lat.min = 31,
  lat.max = 36,
  lon.min = -126,
  lon.max = -119,
  dlat = 5,
  dlon = 5,
  draw.grid = T
)
}
\description{
Test Map of California
}
\author{
Thomas Bryce Kelly
}
