% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.load.R
\name{trim.satellite}
\alias{trim.satellite}
\title{Tirm Satellite Field}
\usage{
trim.satellite(satellite, lon = NULL, lat = NULL)
}
\arguments{
\item{satellite}{a satellite object as returned from read.satellite()}

\item{lon}{a vector of length two indicating the limits desired. By default it is set to the range in longitude values in the satellite object.}

\item{lat}{same as above but for latitude}
}
\description{
Tirm Satellite Field
}
