% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{which.unique}
\alias{which.unique}
\title{Which Unique}
\usage{
which.unique(x)
}
\arguments{
\item{x}{A vector of entries.}
}
\description{
Find the indicies of the unique entries.
}
\author{
Thomas Bryce Kelly
}
