% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.plot.R
\name{add.boot.trendline}
\alias{add.boot.trendline}
\title{Add Bootstrapped Trendline}
\usage{
add.boot.trendline(model, col = "black", lty = 2, lwd = 1, ...)
}
\description{
Add the maximum likelihood trendline to a figured based on a bootstrap estimation.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
