% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.boxplot.box}
\alias{add.boxplot.box}
\title{Add Boxplot Box}
\usage{
add.boxplot.box(
  x,
  y,
  col = "grey",
  border = "black",
  width = 0.7,
  lty = 1,
  lcol = "black",
  lwd = 1,
  xlwd = NULL,
  outliers = T,
  pcol = "black",
  pch = 1,
  cex = 1
)
}
\arguments{
\item{x}{The x locations for the boxs}

\item{y}{The y values that will be plotted.}
}
\description{
Add a box to a boxplot that was initialized by plot.boxplot.
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
