% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.boxplot.box}
\alias{add.boxplot.box}
\title{Add Boxplot Box}
\usage{
add.boxplot.box(x, y, col = "white", border = "black", CI = 0.95,
  width = 0.7, lty = 1, lcol = "black", lwd = 1, xlwd = NULL)
}
\description{
Add a box to a boxplot that was initialized by plot.boxplot.
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
