% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.log.axis}
\alias{add.log.axis}
\title{Add Log Axis}
\usage{
add.log.axis(side = 1, base = 10, col = "black", color.minor = "grey")
}
\description{
A helper function to add log axis to a plot.
}
\author{
Thomas Bryce Kelly
}
\keyword{Axis}
\keyword{Log}
