% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{add.map.bathy}
\alias{add.map.bathy}
\title{Add Map Bathymetry}
\usage{
add.map.bathy(map, bathy, bathy.levels = c(-100, -1000, -2500),
  bathy.col = "darkgrey", bathy.lwd = 1, bathy.lty = 1)
}
\arguments{
\item{map}{a map object as returned by make.map()}

\item{bathy}{a bathymetry dataframe as produced by get.bathy()}

\item{bathy.levels}{a vector of elevations for which to draw contours. NB: depths are negative elevation.}

\item{bathy.col}{the color of the lines to be drawn}

\item{bathy.lwd}{the weight of the lines}

\item{bathy.lty}{the type of line to be drawn}
}
\description{
add bathymetry contours
}
\author{
Thomas Bryce Kelly
}
