% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{add.map.contour}
\alias{add.map.contour}
\title{Add Map Contours}
\usage{
add.map.contour(
  x,
  y,
  z,
  col = "black",
  levels = NULL,
  n = 3,
  labels = TRUE,
  lty = 1,
  lwd = 1
)
}
\description{
Add Map Contours
}
\author{
Thomas Bryce Kelly
}
