% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{add.map.points}
\alias{add.map.points}
\title{Add Map Points}
\usage{
add.map.points(stn.lon, stn.lat, col = "black", cex = 1, pch = 16)
}
\arguments{
\item{stn.lon}{the longitudes of the points to be drawn}

\item{stn.lat}{the latitude of the points to be drawn}

\item{col}{the color of the points}

\item{cex}{the size of point to be drawn}

\item{pch}{the point character to be used}
}
\description{
Add station points to a map
}
\author{
Laura Whitmore
}
