% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{add.shade}
\alias{add.shade}
\title{Add Shaded Box}
\usage{
add.shade(x, col = "grey", border = NA)
}
\description{
A helper function for adding shaded regions to figures (such as in timeseries).
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
