% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.grid.r
\name{build.section.3d}
\alias{build.section.3d}
\title{Build Section (3D)}
\usage{
build.section.3d(x, y, z, zz, xlim = NULL, ylim = NULL, zlim = NULL,
  x.factor = 1, y.factor = 1, z.factor = 1, x.scale = NULL,
  y.scale = NULL, z.scale = NULL, uncertainty = 1e-06,
  neighborhood = 20, field.name = "zz")
}
\description{
Build Section (3D)
}
\author{
Thomas Bryce Kelly
}
