% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.building.R
\name{build.section}
\alias{build.section}
\title{Build Section}
\usage{
build.section(
  x,
  y,
  z,
  lat = NULL,
  lon = NULL,
  gridder = NULL,
  grid = NULL,
  weight = NULL,
  xlim = NULL,
  ylim = NULL,
  x.factor = NULL,
  y.factor = NULL,
  x.scale = NULL,
  y.scale = NULL,
  uncertainty = 1,
  p = 2,
  field.names = NULL,
  nx = 50,
  ny = 50,
  proj = NULL,
  verbose = T
)
}
\arguments{
\item{x}{dimensions (e.g. lat, lon, depth, section distance, time, etc)}

\item{y}{dimensions (e.g. lat, lon, depth, section distance, time, etc)}

\item{z}{signal to be gridded (e.g. T, S, ...)}

\item{gridder}{A function to perform gridding, options gridIDW (default: inverse distance), gridNN (nearest neighbor), gridNNI (natural neighbor) or gridKrige (Krigging)}

\item{xlim}{Limits of the gridding. These are the bounds of the new x-y grid. Default: NULL will set it based on the data + 10%.}

\item{ylim}{Limits of the gridding. These are the bounds of the new x-y grid. Default: NULL will set it based on the data + 10%.}

\item{x.factor}{The relative scale difference between x and y, used to calculate distances. Take into account actual scale AND the relevent scaling of the system (vertical distance tends to be more important than horizontal distance).}

\item{y.factor}{The relative scale difference between x and y, used to calculate distances. Take into account actual scale AND the relevent scaling of the system (vertical distance tends to be more important than horizontal distance).}

\item{x.scale}{The step size in the new x-y grid. By default the scale is set to generate a grid that is 50x50.}

\item{y.scale}{The step size in the new x-y grid. By default the scale is set to generate a grid that is 50x50.}

\item{uncertainty}{= 0: Scaling applied to the distance from the cener of a grid cell to a vertex, used to add a base-line distance to all measurements. 0 = no minimum, 1 = minimum = to half a grid cell.}

\item{nx}{The number of splits to make in the x direction (defaults to 50). Used only if x.scale is not set.}

\item{ny}{The number of splits to make in the y direction (defaults to 50). Used only if y.scale is not set.}

\item{proj}{A gdal projection string such as from make.proj() function. Used to redistribute grid over non-euclidean surfaces like maps.}

\item{field.name}{Sets the name of the new interpolated field. By default the name is 'z1'}
}
\description{
Build Section
}
\author{
Thomas Bryce Kelly
}
\keyword{Gridding}
