% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{calc.boot.r.squared}
\alias{calc.boot.r.squared}
\title{Calcualte Bootstrapped R Squared}
\usage{
calc.boot.r.squared(model)
}
\description{
Calcualte Bootstrapped R Squared
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
