% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.statistics.r
\name{calc.boot.ssr}
\alias{calc.boot.ssr}
\title{Calculate Bootstrapped Sum of Squared Residuals}
\usage{
calc.boot.ssr(model)
}
\description{
A function to calculate the SSR for a series of bootstrapped regressions.
}
\author{
Thomas Bryce Kelly
}
\keyword{Statistics}
