% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.k.reuer}
\alias{calc.k.reuer}
\title{Calculate k (Reuer)}
\usage{
calc.k.reuer(time, wtime, wspeed, mld, SST, teeter.mod = TRUE)
}
\arguments{
\item{time}{the time at which you are calculating k for}

\item{wtime}{a vector of times at which a wind speed is recorded}

\item{wspeed}{a vector of prior windspeeds ordered from most recent to oldest}

\item{mld}{mixed layer depth in meters}

\item{SST}{mixed layer temperature in C}

\item{teeter.mod}{a boolean flag to turn on or off the modified version of Reuer's weighting scheme as outlined in Teeter et al. 2018.}
}
\value{
It will return a list with the individual weights and weighted sum (for dianostics), the k value (that you want, m/d), the instantaneous k (also useful, m/d), k.hist showing the history of k values (diagnostic), and time is time before measurement that each diagnostic is referenced against (for diagnostics, d).
}
\description{
This function calculates the temporal length scale of the NCP measurement, aka a piston velocity (m/d). Essentially it calculates the ventilation based on water column and wind speed over a sequence of wind speed measurements (at 10 m height).
}
\references{
Reuer, M. K., Barnett, B. A., Bender, M. L., Falkowski, P. G., and Hendricks, M. B. (2007). New estimates of Southern Ocean biological production rates from O2/Ar ratios and the triple isotope composition of O2. Deep Sea Research Part I: Oceanographic Research Papers 54, 951–974. doi:10.1016/j.dsr.2007.02.007.

Teeter, L., Hamme, R. C., Ianson, D., and Bianucci, L. (2018). Accurate Estimation of Net Community Production From O 2 /Ar Measurements. Global Biogeochem. Cycles. doi:10.1029/2017GB005874.
}
\author{
Thomas Bryce Kelly
}
\keyword{Air-sea}
\keyword{NCP}
\keyword{Oceanography}
