% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{calc.ptemp}
\alias{calc.ptemp}
\title{Calculate Potential Temperature}
\usage{
calc.ptemp(S, Tmp, P, P.ref, verbose = T)
}
\arguments{
\item{S}{Salinity in PSU}

\item{Tmp}{Temperature in centigrade}

\item{P}{Sample pressure in db}

\item{P.ref}{Reference pressure, typically 0 db (surface).}
}
\description{
Calculate Potential Temperature
}
\author{
Thomas Bryce Kelly
}
