% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.physics.r
\name{conv.cond.to.sal}
\alias{conv.cond.to.sal}
\title{Convert Conductivity to Salinity}
\usage{
conv.cond.to.sal(C = 10000, T = 10, verbose = T)
}
\arguments{
\item{C}{Conductivity in uS cm-1}

\item{T}{Temperature in centigrade}
}
\description{
Convert from conductivity observations to the seawater salinity.
}
\author{
Thomas Bryce Kelly
}
