% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.satellite.r
\name{conv.kahru.chl}
\alias{conv.kahru.chl}
\title{Convert Data to Chl (Kahru)}
\usage{
conv.kahru.chl(x)
}
\arguments{
\item{x}{Raw value}
}
\description{
Convert Kahru's data to mg Chl m-2
}
\author{
Thomas Bryce Kelly
}
