% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{get.pal}
\alias{get.pal}
\title{Get Pal}
\usage{
get.pal(n = 10, pal = "ocean.haline", rev = FALSE)
}
\arguments{
\item{n}{The number of colors desired.}

\item{pal}{A character string of the pallet generating function.}

\item{rev}{A boolean used to flip the order of the colors.}
}
\description{
A helper function to retrieve a vector of colors derived from a given palette generating function.
}
\author{
Thomas Bryce Kelly
}
\keyword{Colormap}
