% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frrf.r
\name{load.frrf}
\alias{load.frrf}
\title{Load FRRF Datafiles}
\usage{
load.frrf(input.dir, file.names = NULL, verbose = T)
}
\arguments{
\item{input.dir}{the input directory}

\item{files.names}{the name of the file in the input directory (default will load all files in that directory)}
}
\description{
Load FRRF Datafiles
}
\author{
Thomas Bryce Kelly
}
