% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.test.R
\name{make.map.arctic}
\alias{make.map.arctic}
\title{Make an Arctic Map}
\usage{
make.map.arctic(
  coast = "coastline2",
  lon = 0,
  lat = 90,
  scale = 2850,
  p = NULL,
  dlat = 10,
  dlon = 60,
  land.col = "lightgray",
  draw.grid = TRUE
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastline2'.}

\item{p}{a projection string, such as those generated by 'make.proj()'}

\item{dlat}{The spacing for the latitude grid (in degrees)}

\item{dlon}{The spacing for the longitude grid (in degrees)}

\item{land.col}{A color string or value for the land polygon}
}
\description{
An example map of the Arctic -- basemap only. For a full test map call test.map.arctic()
}
\author{
Laura Whitmore
}
