% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{make.map.cce}
\alias{make.map.cce}
\title{Make Map (CCE)}
\usage{
make.map.cce(
  coast = "coastlineWorldFine",
  lon.min = -126,
  lon.max = -119,
  lat.min = 31,
  lat.max = 36,
  p = "+proj=merc",
  land.col = "lightgray",
  grid = TRUE,
  dlon = 5,
  dlat = 5
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastlineWorld'.}

\item{lon.min}{The minimum longitude displayed on the map.}

\item{lon.max}{The maximum longitude on the map}

\item{lat.min}{The minimum latitude shown on the map}

\item{lat.max}{The maximum latitude shown on the map}

\item{p}{a projection string, such as those generated by 'make.proj()'}

\item{land.col}{A color string or value for the land polygon}

\item{grid}{Boolean, draw lat/lon grid?}

\item{dlon}{The spacing for the longitude grid (in degrees)}

\item{dlat}{The spacing for the latitude grid (in degrees)}
}
\description{
An example map of the CCE study region.
}
\author{
Thomas Bryce Kelly
}
