% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.test.R
\name{make.map.nga}
\alias{make.map.nga}
\title{Make Map (NGA)}
\usage{
make.map.nga(
  coast = "coastline3",
  lon = -150,
  lat = 58,
  scale = 300,
  p = make.proj("stere", lat = 60, lon = -150),
  dlat = 3,
  dlon = 3,
  land.col = "#333333",
  draw.grid = TRUE
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastline2'.}

\item{p}{a projection string, such as those generated by 'make.proj()'}

\item{dlat}{The spacing for the latitude grid (in degrees)}

\item{dlon}{The spacing for the longitude grid (in degrees)}

\item{land.col}{A color string or value for the land polygon}
}
\description{
An example map of the NGA -- basemap only.
}
\author{
Thomas Bryce Kelly
}
