% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{make.map}
\alias{make.map}
\title{Make Map}
\usage{
make.map(
  coast = NULL,
  lon.min = -75,
  lon.max = 75,
  lat.min = -45,
  lat.max = 45,
  p = NULL,
  land.col = "lightgray",
  grid = TRUE,
  dlon = 15,
  dlat = 15
)
}
\arguments{
\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastlineWorld'.}

\item{lon.min}{The minimum longitude displayed on the map.}
}
\description{
Default settings are for Arctic, lowest resolution coastline.
Exaples of projection arguments include: `+proj=stere +lat_0=90`, `+proj=merc`, `+proj=aea +lat_1=30 +lat_2=45 +lon_0=-120`.
}
\author{
Laura Whitmore

Thomas Bryce Kelly
}
