% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.maps.r
\name{make.proj}
\alias{make.proj}
\title{Make Map Projection}
\usage{
make.proj(projection = NULL, lat = NULL, lon = NULL, h = NULL, dlat = 10)
}
\description{
Make Map Projection
}
\author{
Thomas Bryce Kelly
}
