% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{make.time}
\alias{make.time}
\title{Make Datetime Object}
\usage{
make.time(year = NULL, month = 1, day = 1, hour = 0, minute = 0,
  second = 0, tz = "GMT")
}
\description{
A helper function to generate a datetime object
}
