% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.beta.r
\name{plot.beta.counts}
\alias{plot.beta.counts}
\title{Plot Beta Counts}
\usage{
plot.beta.counts(
  data,
  ylim = NULL,
  cols = pals::glasbey(5),
  show.legend = TRUE
)
}
\arguments{
\item{data}{A beta counter object as returned by _load.beta()_.}

\item{ylim}{The ylim to use. The default uses the limits of the data.}

\item{cols}{A set of 5 colors to use in the plot.}

\item{show.legend}{A boolean flag to turn the legend on and off.}
}
\description{
A helper function to quickly plot up raw data from a beta count data file.
}
\author{
Thomas Bryce Kelly
}
\keyword{Thorium}
