% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.r
\name{plot.boxplot}
\alias{plot.boxplot}
\title{Plot Boxplot}
\usage{
plot.boxplot(
  n = 10,
  ylim = c(0, 1),
  at = NULL,
  main = NULL,
  labels = NULL,
  ylab = NULL,
  xlab = NULL,
  ...
)
}
\arguments{
\item{n}{The}
}
\description{
Generate a blank boxplot
}
\author{
Thomas Bryce Kelly
}
\keyword{Plotting}
