% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.quiver}
\alias{plot.quiver}
\title{Quiver Plot}
\usage{
plot.quiver(
  x,
  y,
  u,
  v,
  scale = NULL,
  xlim = NULL,
  ylim = NULL,
  col = "black",
  ...
)
}
\arguments{
\item{x}{x locations for quivers}

\item{y}{y locations for quivers}

\item{u}{The x component for the quiver arrow}

\item{v}{The y component for the quiver arrow}

\item{scale}{Used to scale the arrow length (optional)}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{col}{the quiver color}

\item{...}{optional arguments passed to plot()}
}
\description{
Quiver Plot
}
\author{
Thomas Bryce Kelly
}
