% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.load.R
\name{read.satellite}
\alias{read.satellite}
\title{Read Satellite Data}
\usage{
read.satellite(file = NULL, entry = 1, verbose = F, lon = NULL, lat = NULL)
}
\description{
read
}
\author{
Thomas Bryce Kelly
}
