% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.load.R
\name{read.satellite}
\alias{read.satellite}
\title{Read Satellite Data}
\usage{
read.satellite(
  file,
  lon = NULL,
  lat = NULL,
  combine = F,
  combine.func = function(x) {
     mean(x, na.rm = T)
 },
  verbose = T
)
}
\description{
read satellite netcdf data from one or more files.
}
\author{
Thomas Bryce Kelly
}
