% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.satellite.r
\name{read.satellite}
\alias{read.satellite}
\title{Read Satellite Data}
\usage{
read.satellite(input.dir, file, conv = function(x) {     x },
  entry = 1)
}
\description{
Read Satellite Data
}
\author{
Thomas Bryce Kelly
}
