% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.satellite.r
\name{read.satellite}
\alias{read.satellite}
\title{Read Satellite Data}
\usage{
read.satellite(
  input.dir = NULL,
  file = NULL,
  conv = function(x) {     x },
  entry = 1,
  verbose = F,
  lon = NULL,
  lat = NULL
)
}
\arguments{
\item{input.dir}{the input directory}
}
\description{
read
}
\author{
Thomas Bryce Kelly
}
